requirements.txt
Flask>=1.1.2
requests>=2.22.0
cloudscraper>=1.2.40
pyzmq>=19.0.1
